Fix disposed DataBehavior asyn behavior
* When a DataBehavior is disposed then the private collection is no
longer valid so any async functions should be aborted.  The best
place to do this is between fetching ids (which is usually already in
process and can take time) and fetching objects.

* Also fixed async references to private collection in cacheMixin so
that any fetches that complete after the private collection has been
disposed do not try to update the non-existent collection.